Circular dependancy removed
